+++ cmd/substreams/devenv.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"github.com/streamingfast/substreams/tools"
+	"github.com/streamingfast/substreams/tools/devenv"
+)
+
+func init() {
+	tools.Cmd.AddCommand(devenv.Cmd)
+}